refactor: restructure dialog overlay, extract logic to mixin
